+++ hazGAN/__archive__/batch.py
