+++ scripts/training/train_minimal.py
@@ -31,4 +31,28 @@ config['nconditions'] = len(metadata['labels'])
 cgan = hazzy.conditional.compile_wgan(config)
 cgan.fit(train, epochs=1, steps_per_epoch=1)
 
+# %% EDA (move later)
+import os
+import xarray as xr
+from environs import Env
+import matplotlib.pyplot as plt
+
+env = Env()
+env.read_env(recurse=True)
+dir = env.str('TRAINDIR')
+
+data = xr.open_dataset(os.path.join(dir, 'data.nc'))
+data['maxwind'] = data['anomaly'].isel(channel=0).max(dim=['lat', 'lon'])
+maxwind = data['maxwind'].data
+plt.hist(maxwind)
+# %%
+data  = data.where(data['maxwind'] > 20, drop=True)
+maxwind = data['maxwind'].data
+plt.hist(maxwind) # this is distribution of maxwinds above 20 mps
+
+# %% sample from this distribution
+import numpy as np
+sample = np.random.choice(maxwind, 100)
+plt.hist(sample)
+
 # %%
